// src/request/dto/request/response-request.dto.ts
export class RequestResponseDto {
  Id: number;
  TicketNumber: string;
  GuestName: string;
  GuestRoom: string;
  HotelId: number;
  CreatorUserId: number;
  AssignedUserId: number;
  RequestTypeId: number;
  RequestStateId: number;
  UserEntryDate: string;
  Comments: string;
  RoomId: number;
  SystemEntryDate: string;
}